Initial bazel project superstructure.
This adds a GRPC project using protoc and io_grpc, with an example project in Armeria.

This has no dependency injection at all, just sets up handlers.

It does add an indirection (manually created) called an Adapter, which allows for a lambda to return the service handler (implementation of the generated AsyncService interface). This will ultimately allow us to do dependency injection with a per-grpc-call sub-graph (in dagger terms, a Subcomponent), with the lambda being the seam.