Begin the Java annotation processing implementation.
This will use the KSP apis, but wrap them, to allow for code reuse for the validation and metadata layer. Wrappers will be minimal and thin, that is, often only partially implementing the KSP APIs, where needed, and allowing the underlying APT types to be re-extracted.

This commit just does this shared infra to start, and sets up an initial annotation processor, which grabs the annotated classes, turns them into parts of the metadata type, and then does nothing.