Add in an initial APT processor which generates a grpc AsyncService adapter. This adapter allows DaggerGrpc to delegate instantiation of the service implementation class to the dagger graph, specifically in a call-scoped subgraph.  Also updates the example to use the processor, and remove the manually constructed example adapters.
Lastly, this actually uses the KotlinSymbolProcessing apis, because it shares validation and metadata construction with the ksp variant. To do this, it uses the ksp-apt bridge in this project.  The generator is very similar to the kotlin/ksp one, but due to the differences between kotlinpoet and javapoet, no attempt to use one system was made here. The apt variant just uses javapoet natively.